requirements.txt
annotated-types==0.7.0
attrs==25.3.0
Automat==24.8.1
beautifulsoup4==4.13.3
blinker==1.9.0
blis==0.7.11
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
confection==0.1.5
constantly==23.10.4
cryptography==44.0.2
cssselect==1.3.0
cymem==2.0.11
defusedxml==0.7.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
filelock==3.18.0
Flask==3.1.0
fuzzywuzzy==0.18.0
greenlet==3.1.1
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
Levenshtein==0.27.1
lxml==5.3.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.12
mysql-connector-python==9.2.0
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parsel==1.10.0
playwright==1.51.0
preshed==3.0.9
Protego==0.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyDispatcher==2.0.7
pyee==12.1.1
Pygments==2.19.1
pyOpenSSL==25.0.0
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-Levenshtein==0.27.1
pytz==2025.1
queuelib==1.7.0
RapidFuzz==3.12.2
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.2
Scrapy==2.12.0
scrapy-playwright==0.0.43
selenium==4.29.0
service-identity==24.2.0
setuptools==76.1.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
thinc==8.2.5
threadpoolctl==3.6.0
tldextract==5.1.3
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
Twisted==24.11.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
w3lib==2.3.1
wasabi==1.1.3
weasel==0.4.1
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
zope.interface==7.2
